Implements automated deployment and Jakefile
The combination of a proc.json properties file, a post-receive hook and
a Jakefile now implement an automated build system.

The Jakefile should be in the specific project folder, and have project
specific tasks. The post-receive hook is installed in the git remote,
in the hooks directory of the bare repository. The proc.json is then
located in whatever setup directory is suitable.

The deployment process can then be reduced to adding the production
server as a git remote, and pushing the desired branch live.

  Ex. git push deploy live:master

This became rather an epic to get sorted, and devulged into a
test-and-push cycle, so I've rebased the history for the last 20 odd
commits. I've listed the oneliners below however.

  - Working on deployment script
  - Renames jakefile -> Jakefile
  - Adds npm install as a prestart hook
  - Adds stream-splitter to repo
  - Changes npm hooks
  - Npm hook modification
  - Bumps stream splitter
  - Reverted package.json
  - Fixes dependency issue in jakefile
  - Removed recurance in package.json
  - Changes to jakefile and proc
  - Fixed exec error
  - Adds prepublish hook
  - Avoids infinite loop
  - Makes fix for removed symlink check
  - Formats proc.json
  - Fixes livePath
  - Changes bower to install into correct dir
  - Prints __dirname
  - Adds forever to package.json
  - Adds final code to start/stop daemon service
  - Refines Jakefile
  - Fixes exec
  - Fixes arguments passed to forever
  - Made paths absolute to start script
  - Adds asset compiling to deploy
  - Removes makefile
  - Moves assets:compile dependency to the correct task
  - Changes assets:compile dep
  - Edits compilation methods
  - Fixes app.js compile issue
  - Forces daemon:stop before start
  - Moves daemon:stop to be a pre-requisite
  - Moved order of arguments for forever
  - Alters daemon
  - Hopefully, the final commit
  - Changes log to append
  - Small test for deployment
  - Changes default port to 50,000