Added courses routes, api calls on fronted and other fixes
Added a route for the api '/api/courses/:year/:id'/ that at the moment
reads a JSON file in the public folder and return it. On the fronted
side, the course controller now gets the data from the api call.
Fixed bower component bug, the params bug and some other small
bugs. #23 #17