Implements first parser test #8 #18 #30
Testing is using mocha, with should.js.

Makes an attempt at setting a test pattern for parsers. Modifies the
timetable parser to a more suitable format, though decisions need to be
made about linking to notes.

My best guess at the moment is that an API that returns bare course
data, along with embedded links to pull notes/exercises will work best
in our suituation.

By forcing the course page to display up-to-date information we prevent
confusion on what should be a reliable record, and also allow us to
populate the database to maintain current CATe links for tags.